+++ src/mapfile_parser/mapfile.py
@@ -660,6 +660,19 @@ class MapFile:
         mapfile.parseMapContentsLLD(mapContents)
         return mapfile
 
+    @staticmethod
+    def newFromMwMapStr(mapContents: str) -> MapFile:
+        """
+        Parses the contents of a Metrowerks ld (mwld) map.
+
+        The `map_contents` argument must contain the contents of a Metrowerks ld mapfile.
+        """
+
+        nativeMapFile = MapFileRs.newFromMwMapStr(mapContents)
+
+        mapfile = MapFile()
+        mapfile._transferContentsFromNativeMapFile(nativeMapFile)
+        return mapfile
 
     def _transferContentsFromNativeMapFile(self, nativeMapFile: MapFileRs):
         for segment in nativeMapFile:
@@ -700,6 +713,7 @@ class MapFile:
         Currently supported map formats:
         - GNU ld
         - clang ld.lld
+        - Metrowerks ld
         """
 
         nativeMapFile = MapFileRs()
@@ -719,6 +733,7 @@ class MapFile:
         Currently supported mapfile formats:
         - GNU ld
         - clang ld.lld
+        - Metrowerks ld
         """
 
         nativeMapFile = MapFileRs()