+++ src/mapfile_parser/frontends/symbol_sizes_csv.py
@@ -41,7 +41,7 @@ def doSymbolSizesCsv(mapPath: Path, outputPath: Path|None, filterSection: str|No
 def processArguments(args: argparse.Namespace, decompConfig=None):
     if decompConfig is not None:
         version = decompConfig.get_version_by_name(args.version)
-        mapPath = Path(args.mapfile if args.mapfile is not None else version.paths.get("map"))
+        mapPath = Path(version.paths.get("map"))
     else:
         mapPath = args.mapfile
 
@@ -56,15 +56,19 @@ def processArguments(args: argparse.Namespace, decompConfig=None):
 def addSubparser(subparser: argparse._SubParsersAction[argparse.ArgumentParser], decompConfig=None):
     parser = subparser.add_parser("symbol_sizes_csv", help="Produces a csv summarizing the files sizes by parsing a map file.")
 
-    nargs: str|int = 1
+    emitMapfile = True
     if decompConfig is not None:
-        nargs = "?"
         versions = []
         for version in decompConfig.versions:
             versions.append(version.name)
-        parser.add_argument("-v", "--version", help="Version to process from the decomp.yaml file", type=str, choices=versions, default=versions[0])
 
-    parser.add_argument("mapfile", help="Path to a map file. This argument is optional if an `decomp.yaml` file is detected on the current project.", type=Path, nargs=nargs)
+        if len(versions) > 0:
+            parser.add_argument("-v", "--version", help="Version to process from the decomp.yaml file", type=str, choices=versions, default=versions[0])
+            emitMapfile = False
+
+    if emitMapfile:
+        parser.add_argument("mapfile", help="Path to a map file.", type=Path)
+
     parser.add_argument("-o", "--output", help="Output path of for the generated csv. If omitted then stdout is used instead.")
     parser.add_argument("--same-folder", help="Mix files in the same folder.", action="store_true")
     parser.add_argument("--symbols", help="Prints the size of every symbol instead of a summary.", action="store_true")